Feature: color when wall is clicked, and color resets with each path
Added feature to color the square red when one in the wall is clicked. 
It resets to black when mouse is moved out of the square.

Resets the colors each time a new path is calculated, so that only the 
last one will be colored.

Also added some more method descriptions.